+++ src/language/ExpressionBuilder.ts
@@ -3,21 +3,19 @@ import {
   ParameterType,
   ParameterTypeRegistry,
 } from '@cucumber/cucumber-expressions'
-import { DocumentUri, LocationLink, Range } from 'vscode-languageserver-types'
 
+import { createLocationLink, makeParameterType, sortLinks, syntaxNode } from './helpers.js'
 import { getLanguage } from './languages.js'
+import { SourceAnalyzer } from './SourceAnalyzer.js'
 import {
   ExpressionBuilderResult,
   ExpressionLink,
+  Language,
   LanguageName,
-  Link,
   ParameterTypeLink,
   ParameterTypeMeta,
   ParserAdapter,
   Source,
-  TreeSitterQueryMatch,
-  TreeSitterSyntaxNode,
-  TreeSitterTree,
 } from './types.js'
 
 export class ExpressionBuilder {
@@ -28,19 +26,9 @@ export class ExpressionBuilder {
     parameterTypes: readonly ParameterTypeMeta[]
   ): ExpressionBuilderResult {
     const expressionLinks: ExpressionLink[] = []
-    const parameterTypeLinks: ParameterTypeLink[] = []
     const errors: Error[] = []
     const registry = new ParameterTypeRegistry()
     const expressionFactory = new ExpressionFactory(registry)
-
-    const treeByContent = new Map<Source<LanguageName>, TreeSitterTree>()
-    const parse = (source: Source<LanguageName>): TreeSitterTree => {
-      let tree: TreeSitterTree | undefined = treeByContent.get(source)
-      if (!tree) {
-        treeByContent.set(source, (tree = this.parserAdapter.parser.parse(source.content)))
-      }
-      return tree
-    }
 
     function defineParameterType(parameterType: ParameterType<unknown>) {
       try {
@@ -54,61 +42,38 @@ export class ExpressionBuilder {
       defineParameterType(makeParameterType(parameterType.name, new RegExp(parameterType.regexp)))
     }
 
-    for (const source of sources) {
-      this.parserAdapter.setLanguageName(source.languageName)
-      let tree: TreeSitterTree
-      try {
-        tree = parse(source)
-      } catch (err) {
-        err.message += `\nuri: ${source.uri}`
-        errors.push(err)
-        continue
-      }
+    const sourceAnalyser = new SourceAnalyzer(this.parserAdapter, sources)
 
-      const language = getLanguage(source.languageName)
-      for (const defineParameterTypeQuery of language.defineParameterTypeQueries) {
-        const query = this.parserAdapter.query(defineParameterTypeQuery)
-        const matches = query.matches(tree.rootNode)
-        for (const match of matches) {
-          const nameNode = syntaxNode(match, 'name')
-          const expressionNode = syntaxNode(match, 'expression')
-          const rootNode = syntaxNode(match, 'root')
-          if (nameNode && expressionNode && rootNode) {
-            const parameterType = makeParameterType(
-              toString(nameNode.text),
-              language.convertParameterTypeExpression(expressionNode.text)
-            )
-            defineParameterType(parameterType)
-            const locationLink = createLocationLink(rootNode, expressionNode, source.uri)
-            parameterTypeLinks.push({ parameterType, locationLink })
-          }
-        }
+    const parameterTypeMatches = sourceAnalyser.getSourceMatches(
+      (language: Language) => language.defineParameterTypeQueries
+    )
+
+    let parameterTypeLinks: ParameterTypeLink[] = []
+    for (const [language, matches] of parameterTypeMatches.entries()) {
+      const links = language.buildParameterTypeLinks(matches)
+      parameterTypeLinks = parameterTypeLinks.concat(links)
+      for (const { parameterType } of links) {
+        defineParameterType(parameterType)
       }
     }
 
-    for (const source of sources) {
-      this.parserAdapter.setLanguageName(source.languageName)
-      const tree = treeByContent.get(source)
-      if (!tree) {
-        continue
-      }
+    const stepDefinitionMatches = sourceAnalyser.getSourceMatches(
+      (language: Language) => language.defineStepDefinitionQueries
+    )
 
-      const language = getLanguage(source.languageName)
-      for (const defineStepDefinitionQuery of language.defineStepDefinitionQueries) {
-        const query = this.parserAdapter.query(defineStepDefinitionQuery)
-        const matches = query.matches(tree.rootNode)
-        for (const match of matches) {
-          const expressionNode = syntaxNode(match, 'expression')
-          const rootNode = syntaxNode(match, 'root')
-          if (expressionNode && rootNode) {
-            const stringOrRegexp = language.convertStepDefinitionExpression(expressionNode.text)
-            try {
-              const expression = expressionFactory.createExpression(stringOrRegexp)
-              const locationLink = createLocationLink(rootNode, expressionNode, source.uri)
-              expressionLinks.push({ expression, locationLink })
-            } catch (err) {
-              errors.push(err)
-            }
+    for (const [, sourceMatches] of stepDefinitionMatches.entries()) {
+      for (const { source, match } of sourceMatches) {
+        const expressionNode = syntaxNode(match, 'expression')
+        const rootNode = syntaxNode(match, 'root')
+        if (expressionNode && rootNode) {
+          const language = getLanguage(source.languageName)
+          const stringOrRegexp = language.convertStepDefinitionExpression(expressionNode.text)
+          try {
+            const expression = expressionFactory.createExpression(stringOrRegexp)
+            const locationLink = createLocationLink(rootNode, expressionNode, source.uri)
+            expressionLinks.push({ expression, locationLink })
+          } catch (err) {
+            errors.push(err)
           }
         }
       }
@@ -117,55 +82,8 @@ export class ExpressionBuilder {
     return {
       expressionLinks: sortLinks(expressionLinks),
       parameterTypeLinks: sortLinks(parameterTypeLinks),
-      errors,
+      errors: sourceAnalyser.getErrors().concat(errors),
       registry,
     }
   }
-}
-
-function toString(s: string): string {
-  const match = s.match(/^['"](.*)['"]$/)
-  if (!match) return s
-  return match[1]
-}
-
-function syntaxNode(match: TreeSitterQueryMatch, name: string): TreeSitterSyntaxNode | undefined {
-  return match.captures.find((c) => c.name === name)?.node
-}
-
-function makeParameterType(name: string, regexp: string | RegExp) {
-  return new ParameterType(name, regexp, Object, (arg) => arg, true, false)
-}
-
-function sortLinks<L extends Link>(links: L[]): readonly L[] {
-  return links.sort((a, b) => {
-    const pathComparison = a.locationLink.targetUri.localeCompare(b.locationLink.targetUri)
-    if (pathComparison !== 0) return pathComparison
-    return a.locationLink.targetRange.start.line - b.locationLink.targetRange.start.line
-  })
-}
-
-function createLocationLink(
-  rootNode: TreeSitterSyntaxNode,
-  expressionNode: TreeSitterSyntaxNode,
-  targetUri: DocumentUri
-) {
-  const targetRange: Range = Range.create(
-    rootNode.startPosition.row,
-    rootNode.startPosition.column,
-    rootNode.endPosition.row,
-    rootNode.endPosition.column
-  )
-  const targetSelectionRange: Range = Range.create(
-    expressionNode.startPosition.row,
-    expressionNode.startPosition.column,
-    expressionNode.endPosition.row,
-    expressionNode.endPosition.column
-  )
-  const locationLink: LocationLink = {
-    targetRange,
-    targetSelectionRange,
-    targetUri,
-  }
-  return locationLink
 }