+++ test/language/ExpressionBuilder.test.ts
@@ -9,7 +9,7 @@ import { ParserAdapter, Source } from '../../src/language/types.js'
 import { NodeParserAdapter } from '../../src/tree-sitter-node/NodeParserAdapter.js'
 import { WasmParserAdapter } from '../../src/tree-sitter-wasm/WasmParserAdapter.js'
 
-const parameterTypeSupport: Set<LanguageName> = new Set(['typescript', 'java', 'ruby'])
+const parameterTypeSupport: Set<LanguageName> = new Set(['typescript', 'java', 'ruby', 'c_sharp'])
 
 function defineContract(makeParserAdapter: () => ParserAdapter) {
   let expressionBuilder: ExpressionBuilder
@@ -21,19 +21,31 @@ function defineContract(makeParserAdapter: () => ParserAdapter) {
 
   for (const dir of glob.sync(`test/language/testdata/*`)) {
     const languageName = basename(dir) as LanguageName
-    // if (languageName !== 'php') {
-    //   continue
-    // }
+
+    if (languageName === 'c_sharp') {
+      it(`builds parameter type from [StepArgumentTransformation] without expression`, async () => {
+        const sources = await loadSources(dir, languageName)
+        const result = expressionBuilder.build(sources, [{ regexp: '.*', name: 'int' }])
+
+        const regexpStrings = result.parameterTypeLinks.find(
+          (l) => l.parameterType.name === 'WithoutExpression'
+        )?.parameterType?.regexpStrings
+        assert.deepStrictEqual(regexpStrings, ['.*'])
+      })
+
+      it(`builds parameter type from multiple [StepArgumentTransformation] with the same return type`, async () => {
+        const sources = await loadSources(dir, languageName)
+        const result = expressionBuilder.build(sources, [{ regexp: '.*', name: 'int' }])
+
+        const regexpStrings = result.parameterTypeLinks.find(
+          (l) => l.parameterType.name === 'DateTime'
+        )?.parameterType?.regexpStrings
+        assert.deepStrictEqual(regexpStrings, ['today', 'tomorrow', '(.*) days later'])
+      })
+    }
+
     it(`builds parameter types and expressions from ${languageName} source`, async () => {
-      const sources: Source<LanguageName>[] = await Promise.all(
-        glob.sync(`${dir}/**/*`).map((path) => {
-          return readFile(path, 'utf-8').then((content) => ({
-            languageName,
-            content,
-            uri: `file://${resolve(path)}`,
-          }))
-        })
-      )
+      const sources = await loadSources(dir, languageName)
       const result = expressionBuilder.build(sources, [{ regexp: '.*', name: 'int' }])
 
       // verify that the targetSelectionRange is inside the targetRange
@@ -82,6 +94,21 @@ Please register a ParameterType for 'undefined-parameter'`,
       }
     })
   }
+}
+
+async function loadSources(
+  dir: string,
+  languageName: LanguageName
+): Promise<Source<LanguageName>[]> {
+  return Promise.all(
+    glob.sync(`${dir}/**/*`).map((path) => {
+      return readFile(path, 'utf-8').then((content) => ({
+        languageName,
+        content,
+        uri: `file://${resolve(path)}`,
+      }))
+    })
+  )
 }
 
 describe('ExpressionBuilder', () => {